One of the tests should stay in the dev_env_tests because we dont get a fixture for controller
